Make the tests compatible with Python 2.6.
Had to rely on unittest2 due to all the new features added in 2.7.  Also had
to work around a wacky bug in 2.6, where a context manager's __exit__ receives
an exception's arguments rather than the exception object, for a few cases of
built-in exceptions.

Only change necessary to pathlib itself was to replace {} with {0} in a couple
format strings.